+++ app/models/__init__.py
@@ -3,3 +3,4 @@ from .finalidade import Finalidade
 from .upload_file import UploadFile
 from .equipe import Equipe
 from .circulo import Circulo
+from .encontreiro import Encontreiro